Extract weaver validation into reusable weavercheck package
